таймер активности пользователя

+++ utils/helpers.py
@@ -5,6 +5,9 @@ from aiogram import types
 async def delete_message_after_delay(message: types.Message, delay: int):
     await asyncio.sleep(delay)
     try:
-        await message.delete()
+        if message:
+            await message.delete()
+        else:
+            return
     except Exception as e:
         logging.info(f"Could not delete message: {e}")
